getting the real data and using it the searchbar component
